refactor(866): promote three cross-section reads onto IUserServiceRead
Read-boundary work ahead of the section move: five out-of-section consumers were
injecting the full IUserService while only reading.

Three read members move up from IUserService to IUserServiceRead, along with
UserParticipationRow (all DTO/Guid-returning, so HUM0029-clean):
- GetByEmailOrAlternateAsync    — GoogleIntegration x3
- GetAllParticipationsForYearAsync — Tickets x2
- GetAccountsDueForAnonymizationAsync — ProcessAccountDeletionsJob

[SurfaceBudget(6)] -> (9) on IUserServiceRead follows mechanically.

Demoted to IUserServiceRead: Shifts/ShiftProfileController, Teams/TeamService,
Tickets/TicketQueryService, Infrastructure's TermRenewalReminderJob and
ProcessAccountDeletionsJob.

Two Teams test locators registered the substitute only under IUserService;
they now register IUserServiceRead as well, matching production DI, which
resolves CachingUserService under both.

The sixteen consumers that genuinely write through IUserService are untouched
and still hold the full interface.

Refs nobodies-collective/Humans#866
