+++ src/Humans.Infrastructure/Jobs/ProcessAccountDeletionsJob.cs
@@ -5,6 +5,7 @@ using Humans.Application.Interfaces;
 using Humans.AuditLog.Contracts;
 using Humans.Email.Contracts;
 using Humans.Application.Interfaces.Users;
+using Humans.Users.Contracts;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
 
@@ -28,7 +29,7 @@ namespace Humans.Infrastructure.Jobs;
 /// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class ProcessAccountDeletionsJob(
-    IUserService userService,
+    IUserServiceRead userService,
     IAccountDeletionService accountDeletionService,
     IEmailService emailService,
     IEmailMessageFactory emailMessages,
