+++ tests/Humans.Teams.Tests/Services/TeamRoleServiceTests.cs
@@ -28,6 +28,7 @@ using Humans.Application.Interfaces.Auth;
 
 using Humans.GoogleIntegration.Data;
 using Humans.Shifts.Data;
+using Humans.Users.Contracts;
 
 namespace Humans.Teams.Tests;
 
@@ -63,6 +64,7 @@ public sealed class TeamRoleServiceTests : TeamsTestHarness
             .With<IGoogleSyncOutboxService>(googleOutboxService)
             .With(teamResourceService)
             .With(userService)
+            .With<IUserServiceRead>(userService)
             .Build();
         var shiftManagementService = new ShiftManagementService(
             new ShiftRepository(ShiftsDbFactory, ShiftsDb, Clock),
