+++ tests/Humans.Tickets.Tests/Services/TicketSyncServiceTests.cs
@@ -18,6 +18,7 @@ using NodaTime;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
 using Humans.Tickets.Domain;
+using Humans.Users.Contracts;
 
 namespace Humans.Tickets.Tests.Services;
